requirements.txt
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
google-auth==2.18.1
google-auth-oauthlib==1.0.0
gspread==5.9.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
Werkzeug==2.3.4
zipp==3.15.0
